// src/pages/project-details/components/project-models/model-container/model-card/model-card.interface.ts
// INTEL CONFIDENTIAL
//
// Copyright (C) 2021 Intel Corporation
//
// This software and the related documents are Intel copyrighted materials, and your use of them is governed by
// the express license under which they were provided to you ("License"). Unless the License provides otherwise,
// you may not use, modify, copy, publish, distribute, disclose or transmit this software or the related documents
// without Intel's prior written permission.
//
// This software and the related documents are provided as is, with no express or implied warranties,
// other than those that are expressly stated in the License.

import { Performance } from '../../../../../../core/projects';

export interface ModelCardProps {
    id: string;
    accuracy: number;
    performance: Performance;
    version: number;
    isActiveModel: boolean;
    upToDate: boolean;
    creationDate: string;
    architectureId: string;
    architectureName: string;
}